% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{benchmark_TTS}
\alias{benchmark_TTS}
\title{Benchmark Segmentation Performance}
\usage{
benchmark_TTS(lasdf, ...)
}
\arguments{
\item{lasdf}{Point cloud data}

\item{...}{Parameters passed to TTS_segmentation()}
}
\value{
Benchmark results with timing information
}
\description{
Benchmark Segmentation Performance
}
