% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{build_gradient_flow_ultra}
\alias{build_gradient_flow_ultra}
\title{Build Ultra-Fast Gradient Flow}
\usage{
build_gradient_flow_ultra(
  morse_complex,
  vertices,
  max_distance = 2,
  grid_size = 5
)
}
\description{
Build Ultra-Fast Gradient Flow
}
\keyword{internal}
