% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{build_gradient_flow_ultra_paper}
\alias{build_gradient_flow_ultra_paper}
\title{Build Ultra-Fast Gradient Flow (Using correct critical simplices)}
\usage{
build_gradient_flow_ultra_paper(
  morse_complex,
  vertices,
  max_distance = 2,
  grid_size = 5
)
}
\description{
Build Ultra-Fast Gradient Flow (Using correct critical simplices)
}
\keyword{internal}
