% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{build_minima_connectivity_from_saddles}
\alias{build_minima_connectivity_from_saddles}
\title{Build Minima Connectivity from 1-Saddles (R Wrapper)}
\usage{
build_minima_connectivity_from_saddles(morse_complex, minima)
}
\description{
Build Minima Connectivity from 1-Saddles (R Wrapper)
}
\keyword{internal}
