% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{create_segmentation_proper}
\alias{create_segmentation_proper}
\title{Create Final Segmentation}
\usage{
create_segmentation_proper(vertices, gradient_flow, labeled_data)
}
\description{
Create Final Segmentation
}
\keyword{internal}
