% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{detect_tree_seeds_proper}
\alias{detect_tree_seeds_proper}
\title{Detect Tree Seeds from Morse Minima}
\usage{
detect_tree_seeds_proper(minima, vertices, clip_height)
}
\description{
Detect Tree Seeds from Morse Minima
}
\keyword{internal}
