% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{get_CCMESH}
\alias{get_CCMESH}
\title{Ultra-fast mesh preparation with connected components}
\usage{
get_CCMESH(vertices, faces, select_largest = TRUE)
}
\arguments{
\item{vertices}{Vertex data from alphahull}

\item{faces}{Face data from alphahull}

\item{select_largest}{If TRUE, returns only largest connected component (default: TRUE).
If FALSE, returns list of all connected components.}
}
\value{
Single mesh (if select_largest=TRUE) or list of meshes (if select_largest=FALSE)
}
\description{
Ultra-fast mesh preparation with connected components
}
