% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{print.TTS_result}
\alias{print.TTS_result}
\title{Print method for TTS_result}
\usage{
\method{print}{TTS_result}(x, ...)
}
\arguments{
\item{x}{A TTS_result object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for TTS_result
}
