% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{print_TTS_result}
\alias{print_TTS_result}
\title{Print Method for TTS_result}
\usage{
\method{print}{TTS_result}(x, ...)
}
\arguments{
\item{x}{TTS_result object}

\item{...}{Additional parameters}
}
\description{
Print Method for TTS_result
}
