% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{proc_lowerSTAR}
\alias{proc_lowerSTAR}
\title{Process lower star to get Morse pairings}
\usage{
proc_lowerSTAR(list_lowerSTAR, vertex)
}
\arguments{
\item{list_lowerSTAR}{Lower star data}

\item{vertex}{Vertex data}
}
\value{
List with vector field and critical simplices
}
\description{
Process lower star to get Morse pairings
}
\keyword{internal}
