% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{propagate_labels_constrained}
\alias{propagate_labels_constrained}
\title{Propagate Labels via Gradient Flow}
\usage{
propagate_labels_constrained(
  gradient_flow,
  seeds,
  vertices,
  spatial_threshold = 2
)
}
\description{
Propagate Labels via Gradient Flow
}
\keyword{internal}
