% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{propagate_labels_simple_fixed}
\alias{propagate_labels_simple_fixed}
\title{Propagate Labels via Gradient Flow}
\usage{
propagate_labels_simple_fixed(gradient_flow, seeds, vertices)
}
\description{
Propagate Labels via Gradient Flow
}
\keyword{internal}
