% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{save_MORSE_2d}
\alias{save_MORSE_2d}
\title{Save 2D visualization to file}
\usage{
save_MORSE_2d(
  morse_complex,
  filename,
  width = 10,
  height = 8,
  dpi = 300,
  panel_2d = FALSE,
  ...
)
}
\arguments{
\item{morse_complex}{Output from compute_MORSE_complex()}

\item{filename}{Output file name}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Resolution}

\item{panel_2d}{If TRUE, saves multi-panel plot. If FALSE, saves single projection plot.}

\item{...}{Additional arguments to visualize_MORSE_2d() or visualize_MORSE_2d_panel()}
}
\description{
Save 2D visualization to file
}
