% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\docType{package}
\name{TopTreeSegR}
\alias{TopTreeSegR}
\title{Topological Tree Segmentation using Discrete Morse Theory}
\description{
Ultra-fast topology-based tree segmentation from LiDAR point clouds
using Discrete Morse Theory and RcppArmadillo for maximum performance.
}
