% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{validate_segmentation}
\alias{validate_segmentation}
\title{Validate Segmentation Quality}
\usage{
validate_segmentation(TTS_result)
}
\arguments{
\item{TTS_result}{Result from TTS_segmentation()}
}
\value{
Validation metrics
}
\description{
Validate Segmentation Quality
}
