% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{validate_TTS}
\alias{validate_TTS}
\title{' Validate Segmentation Quality}
\usage{
validate_TTS(TTS_result, lasdf, input_truth = "pid", val_col = "treeid")
}
\arguments{
\item{TTS_result}{Result from TTS_segmentation()}

\item{lasdf}{Input point cloud las/laz format}

\item{input_truth}{Attribute from lasdf (default: 'pid')}

\item{val_col}{Column/field in the LAS data storing true labels}
}
\value{
Validation metrics including precision, recall, F1, accuracy, RI, ARI
}
\description{
' Validate Segmentation Quality
}
