% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{plot_TTS_3d}
\alias{plot_TTS_3d}
\title{3D Visualization of Tree Segmentation}
\usage{
plot_TTS_3d(TTS_result, colors = "viridis", point_size = 2, alpha = 0.8)
}
\arguments{
\item{TTS_result}{Result from TTS_segmentation()}

\item{colors}{Color of labeled points (default: "viridis")}

\item{point_size}{Size of points in 3D plot (default: 2)}

\item{alpha}{Point transparency (default: 0.8)}
}
\value{
plotly 3D scatter plot
}
\description{
3D Visualization of Tree Segmentation
}
