% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{print.TTS_pipeline_result}
\alias{print.TTS_pipeline_result}
\title{Print Method for Pipeline Results}
\usage{
\method{print}{TTS_pipeline_result}(x, ...)
}
\description{
Print Method for Pipeline Results
}
