% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTreeSegR.R
\name{validate_TTS}
\alias{validate_TTS}
\title{Validate TTS Segmentation Results}
\usage{
validate_TTS(TTS_result, lasdf, val_col = "treeid", auto_align = TRUE)
}
\arguments{
\item{TTS_result}{Result from TTS_segmentation() with $labels and $original_pid}

\item{lasdf}{Original LAS object with ground truth}

\item{val_col}{Column in LAS with ground truth labels (default: 'treeid')}

\item{auto_align}{Label (true and pridicted) alighment (default: TRUE)}
}
\value{
Validation metrics
}
\description{
Uses result$labels and result$original_pid to match with LAS ground truth
}
